chore: refactor to `ResourceAssistantChat`
